==> notebooks/01_get_company_cik.ipynb <==
Download CIK numbers for companies.
S&P 500 companies were used because they are likely to report on risk factors than smaller/medium sized companies.

import pandas as pd

def get_company_ciks():
    """Get CIK numbers for S&P 500 companies"""
    sp500_url = "https://en.wikipedia.org/wiki/List_of_S%26P_500_companies"
    tables = pd.read_html(sp500_url, storage_options={"User-Agent": "Mozilla/5.0"})
    sp500_df = tables[0]
    
    sp500_df['CIK'] = sp500_df['CIK'].astype(str)
    
    # choosing only these columns
    return sp500_df[['Symbol', 'Security', 'CIK']]


companies = get_company_ciks()

# creating txt 
companies['CIK'].head(50).to_csv("../cik.txt", index=False, header=None)
