==> notebooks/04_baseline_models.ipynb <==
import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.preprocessing import MultiLabelBinarizer
from sklearn.model_selection import train_test_split
from sklearn.metrics import f1_score
from pathlib import Path
import spacy
import re
RSEED = 182026

link = Path(__name__).resolve().parent.parent
df = pd.read_csv(link / 'data/risk_paragraph_labeled.csv')
df.info()

df['year'].value_counts()

#some text contains "•" and other characters that should be removed
contains_dot_df = df[df['paragraph'].str.contains("•")]
len(contains_dot_df)

# checking example of sentences that has "•"
nlp = spacy.load('en_core_web_sm')
doc = nlp(contains_dot_df.iloc[0,1])
for sent in doc.sents:
    print(f"\n{sent}")

#function to take care of such characters
def clean_text(text: str) -> str:
    text = re.sub(r'[•+▪■●◦·]', '', text)
    text = re.sub(r'(?m)^[\-\*]\s*', '', text)  # leading - or *
    text = re.sub(r'\s+', ' ', text)
    return text.strip()

#testing function
cleaned_text = clean_text(contains_dot_df.iloc[0,1])
for sent in nlp(cleaned_text).sents:
    print(f"\n{sent}")

#use clean_text function on paragraph column
df['paragraph'] = df['paragraph'].apply(clean_text)

#check if column still contains such characters
df[df['paragraph'].str.contains("•")]

df.head()

#shape of df
df.shape

#checking duplicates
len(df.drop_duplicates('paragraph'))

# no duplicates found yet. 423 is the same number of rows in df

#sorting columns to check if similar rows appear
df.sort_values(['labels',"cik","paragraph"])

#index 36 and 52 below are identical
#lets investigate

possible_duplicated_paragraphs = df[df['cik']==1018724].sort_values(['labels',"cik","paragraph"])
possible_duplicated_paragraphs.head(10)
#it seems that the same/similar risk paragraphs from previous year (2022) were used for 2023

#showing 2023 and 2022 paragraphs for cik-1018724

for loc in [0,1]:
    for sent in nlp(possible_duplicated_paragraphs.iloc[loc,1]).sents:
        print(f"{sent}")
    print(50*"*")
    print(50*"*")

# due to possible duplicates in the paragraph columns and this project being a "prototype" I will use 
# only 2022 paragraphs

#use 2022 year
df_copy = df[df['year']==2022].copy()
df_copy.info()

df_copy.head()

#function that removes stop words and characters that wont add value to model
def extra_clean(text:str):
    whole_text = []
    doc = nlp(text)
    for token in doc:
        if token.like_num or token.is_punct or token.is_digit or token.is_currency or token.is_space or token.is_stop:
            continue
        lemma = token.lemma_.lower() #using the lemma of the word
        whole_text.append(lemma)
    return " ".join(whole_text)



print(f"original paragraph: {df_copy.iloc[-2]['paragraph']}")

print(f"Cleaned paragraph: {extra_clean(df_copy.iloc[-2]['paragraph'])}")

# words like "the", "some", "of" etc have been removed to 
# ensure the reduction of noise for better performance

#apply function to whole paragraph column
df_copy['paragraph_cleaned'] = df_copy['paragraph'].apply(extra_clean)
df_copy.head()

#save dataset to use.
df_copy.to_csv(link / "data/risk_data.csv", index=False)

#since the labelling is a multi-label classification I could use a multilabelbinarizer
#to transform the label column.

#this requires putting labels in a list
df_copy['label_list'] = df_copy['labels'].apply(lambda x: [l.strip() for l in x.split(',') if l.strip()])

df_copy.head(3)

df_copy["label_list"].value_counts()

#removing Ambiguous labels and rows without labels (if there are any)
df_copy = df_copy[df_copy['label_list'].apply(lambda x: len(x) > 0 and 'Ambiguous' not in x)]

# input and label data
X = df_copy['paragraph_cleaned'].values
y_list = df_copy['label_list'].tolist()

#applying multilabelbinarizer on the label
mlb = MultiLabelBinarizer()
y = mlb.fit_transform(y_list)

#classes for Ml model
mlb.classes_

#vectorizer
vectorizer = TfidfVectorizer(max_features=5000,ngram_range=(1, 2), min_df=2,)

# Split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=RSEED)

#transform paragraphs
X_train_vec = vectorizer.fit_transform(X_train)

# Train one model per class. 
# example: train whether a paragraph is a credit risk or not. same analogy for the rest
#each model will be stored in a dictionary.
all_models = {}
for i, class_name in enumerate(mlb.classes_):
    print(f"Training {class_name}...")
    model = LogisticRegression(
        max_iter=1000,
        class_weight='balanced',
        C=1.0
    )
    model.fit(X_train_vec, y_train[:, i])
    all_models[class_name] = model

all_models

#predicting X_train
y_pred = np.zeros((len(X_train), len(mlb.classes_)))
for i, class_name in enumerate(mlb.classes_):
    probs = all_models[class_name].predict_proba(X_train_vec)[:, 1]
    y_pred[:, i] = (probs >= 0.5).astype(int)

# Overall metrics
f1_micro = f1_score(y_train, y_pred, average='micro')
f1_macro = f1_score(y_train, y_pred, average='macro')
f1_samples = f1_score(y_train, y_pred, average='samples')
print(f"\nOverall Training Metrics:")
print(f"F1 (micro): {f1_micro:.3f}")
print(f"F1 (macro): {f1_macro:.3f}")
print(f"F1 (samples): {f1_samples:.3f}")

#predicting X_test
X_test_vec = vectorizer.transform(X_test)
y_pred_test = np.zeros((len(X_test), len(mlb.classes_)))
for i, class_name in enumerate(mlb.classes_):
    probs = all_models[class_name].predict_proba(X_test_vec)[:, 1]
    y_pred_test[:, i] = (probs >= 0.5).astype(int)


# Overall metrics
f1_micro = f1_score(y_test, y_pred_test, average='micro')
f1_macro = f1_score(y_test, y_pred_test, average='macro')
f1_samples = f1_score(y_test, y_pred_test, average='samples')
print(f"\nOverall Test Metrics:")
print(f"F1 (micro): {f1_micro:.3f}")
print(f"F1 (macro): {f1_macro:.3f}")
print(f"F1 (samples): {f1_samples:.3f}")

#with the ideas shown here a script will be devloped for Training the models
# Logistic regression and xgboost will follow similiar logic

X.shape
